style(web): hide HeaderDropdown on md & moved all landing-page-related to _components at (home)
