+++ apps/web/src/app/(home)/_components/Banner.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { Button } from "../ui/button"
+import { Button } from "../../../components/ui/button"
 
 export default function Banner() {
     const imgUrl = 'https://images.pexels.com/photos/3568518/pexels-photo-3568518.jpeg'
