+++ apps/web/src/app/(home)/_components/Campaign.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { Button } from "../ui/button"
+import { Button } from "../../../components/ui/button"
 
 
 export default function Campaign() {
