+++ apps/web/src/app/(home)/_components/HeroCarousel/HeroCarousel.tsx
